requirements.txt
Django>=5.2
djangorestframework>=3.16
djangorestframework-simplejwt>=5.5
django-cors-headers>=4.9
Pillow>=10.0
scikit-learn>=1.8
numpy>=1.26
pandas>=2.0
django-crispy-forms>=2.6
crispy-bootstrap5>=2026.3
django-filter>=25.0
pymongo>=4.6
requests>=2.31.0
python-dotenv>=1.0.1
